+++ common/src/main/java/net/mehvahdjukaar/amendments/mixins/WallHangingSignBlockMixin.java
@@ -75,7 +75,7 @@ public abstract class WallHangingSignBlockMixin extends Block implements EntityB
         super.entityInside(state, level, pos, entity);
         if (level.isClientSide && ClientConfigs.SWINGING_SIGNS.get() &&
                 level.getBlockEntity(pos) instanceof ExtendedHangingSign tile && tile.getExtension().canSwing()) {
-            tile.getExtension().animation.hitByEntity(entity, state, pos);
+            tile.getExtension().getClientAnimation().hitByEntity(entity, state, pos);
         }
     }